% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{obs2lia}
\alias{obs2lia}
\title{Convert observed heritability to the liability scale}
\usage{
obs2lia(h2 = NULL, K = NULL, P = NULL)
}
\arguments{
\item{h2}{observed SNP heritability}

\item{K}{Population disease prevalence}

\item{P}{Proportion of cases in the sample}
}
\description{
Whether the estiamted heritability is based on GREML or on summary statistics, if the trait is a disease trait,
the estimate may be biased due to assertainment bias. To account for this, one can convert the observed heritability
to the heritability on the liability scale by correcting for population disease prevalence and the proportion of cases in the sample
}
\author{
Palle Duun Rohde
}
