% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2_gwas_func.R
\name{plink_reg}
\alias{plink_reg}
\title{Performs linear/logistic regression using PLINK 2.0 alpha}
\usage{
plink_reg(
  bedfile = NULL,
  bimfile = NULL,
  famfile = NULL,
  yFile = NULL,
  covFile = NULL,
  covar = NULL,
  fids = NULL,
  ids = NULL,
  rsids = NULL,
  fnSNPs = NULL,
  fnOut = NULL,
  wd = NULL,
  logistic = FALSE,
  firth = FALSE,
  ncores = 1,
  dir = NULL
)
}
\arguments{
\item{bedfile}{Path to the PLINK bedfile}

\item{bimfile}{Path to the PLINK bimfile}

\item{famfile}{Path to the PLINK famfile}

\item{yFile}{path to phenotype file, which is a matrix containing fids, ids and the phenotypic values}

\item{covFile}{path to the covariate file, which is a matrix containing fids, ids, and the covariates used in the analysis}

\item{covar}{a character string of covaraites used in the analysis}

\item{fids}{family IDs}

\item{ids}{IDs}

\item{rsids}{vector of SNPs used in the association analysis}

\item{fnSNPs}{path to a list of SNPs if they already are stored on disk}

\item{fnOut}{path and file name of the association results}

\item{wd}{working directory of temporary files}

\item{logistic}{logical if TRUE the logic link function is used}

\item{firth}{logical if TRUE Firth's logistic regression will be performed.}

\item{ncores}{number of cores used in the analysis}

\item{dir}{path to PLINK 2.0 execuatble}
}
\description{
Performs linear/logistic regression using PLINK 2.0 alpha
}
